# services/authentication/swagger/swagger.yaml
---
swagger: "2.0"
info:
  description: "This is a sample API server.  See source code in [Python Microservices](https://github.com/python-microservices/microservices-scaffold)"
  version: "1.0.0"
  title: "API of Films and actors"
  termsOfService: "https://github.com/python-microservices/microservices-scaffold/"
  contact:
    email: "a.vara.1986@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
tags:
- name: "Auth"
  description: "Auth operation with users"

schemes:
- "http"

paths:
  /login:
    post:
      tags:
        - "Auth"
      summary: "Example endpoint for create token"
      description: ""
      operationId: "login"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "payload"
        description: "Data of user for login"
        required: true
        schema:
          $ref: "#/definitions/UserLogin"
      responses:
        200:
          description: "Token"
          schema:
            $ref: '#/definitions/User'
        405:
          description: "Invalid input"
      x-swagger-router-controller: "src.views.auth"

  /auth/verify-token:
    post:
      tags:
        - "Auth"
      summary: "Example endpoint for verify token"
      description: ""
      operationId: "verify_token"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "payload"
        description: "Data of user for login"
        required: true
        schema:
          $ref: "#/definitions/UserVerifyToken"
      responses:
        200:
          description: "Token"
          schema:
            $ref: '#/definitions/User'
        405:
          description: "Invalid input"
      x-swagger-router-controller: "src.views.auth"


definitions:
      UserLogin:
        type: "object"
        properties:
          username:
            type: "string"
          password:
            type: "string"
      UserVerifyToken:
        type: "object"
        properties:
          access_token:
            type: "string"
      User:
        type: "object"
        properties:
          access_token:
            type: "string"

externalDocs:
  description: "Find out more about Swagger"
  url: "https://github.com/python-microservices/microservices-scaffold"